Tidy up

+++ Sources/TMDb/Models/BackdropURLProviding.swift
@@ -1,13 +1,16 @@
 import Foundation
 
+/// Provides URLs for the different sizes of backdrop images.
 public protocol BackdropURLProviding {
 
+    /// Path to the backdrop image.
     var backdropPath: URL? { get }
 
 }
 
 public extension BackdropURLProviding {
 
+    /// Original backdrop image URL.
     var backdropOriginalURL: URL? {
         guard let backdropPath = backdropPath else {
             return nil
@@ -18,6 +21,7 @@ public extension BackdropURLProviding {
             .appendingPathComponent(backdropPath.absoluteString)
     }
 
+    /// Large backdrop image URL.
     var backdropLargeURL: URL? {
         guard let backdropPath = backdropPath else {
             return nil
@@ -28,6 +32,7 @@ public extension BackdropURLProviding {
             .appendingPathComponent(backdropPath.absoluteString)
     }
 
+    /// Medium backdrop image URL.
     var backdropMediumURL: URL? {
         guard let backdropPath = backdropPath else {
             return nil
@@ -38,6 +43,7 @@ public extension BackdropURLProviding {
             .appendingPathComponent(backdropPath.absoluteString)
     }
 
+    /// Small backdrop image URL.
     var backdropSmallURL: URL? {
         guard let backdropPath = backdropPath else {
             return nil
@@ -48,8 +54,9 @@ public extension BackdropURLProviding {
             .appendingPathComponent(backdropPath.absoluteString)
     }
 
-    static var backdropAspectRatio: Float {
-        500 / 281
+    /// Aspect ratio of the backdrop image.
+    static var backdropAspectRatio: Double {
+        500.0 / 281.0
     }
 
 }
